Updating Theme Generator widgets to typescript
