+++ website/package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@dxc-technology/halstack-react": "0.0.0-824a373",
+    "@radix-ui/react-popover": "0.1.6",
     "@types/styled-components": "^5.1.18",
     "axios": "^0.27.2",
     "cross-env": "^7.0.3",
@@ -26,6 +27,7 @@
   "devDependencies": {
     "@types/node": "17.0.2",
     "@types/react": "17.0.37",
+    "@types/react-color": "^3.0.6",
     "eslint": "8.5.0",
     "eslint-config-next": "12.0.9",
     "typescript": "4.5.4"
