+++ website/screens/theme-generator/components/widgets/AlphaValueInput.tsx
@@ -1,21 +1,23 @@
+import { DxcFlex } from "@dxc-technology/halstack-react";
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
+import ThemeInputWidgetProps from "./types";
 
 const AlphaValueInput = ({
   propertyName,
   propertyValue,
   onChangeCustomTheme,
-}) => {
+}: ThemeInputWidgetProps): JSX.Element => {
   const [value, changeValue] = useState(
-    propertyValue.match(/-?[0-9]+(.[0-9]+)?/g)
+    propertyValue.match(/-?[0-9]+(.[0-9]+)?/g)?.join("")
   );
 
   useEffect(() => {
-    changeValue(propertyValue.match(/-?[0-9]+(.[0-9]+)?/g));
+    changeValue(propertyValue.match(/-?[0-9]+(.[0-9]+)?/g)?.join(""));
   }, [propertyValue]);
 
   return (
-    <AlphaValueInputContainer>
+    <DxcFlex alignItems="center">
       <StyledInput
         type="number"
         value={value}
@@ -28,18 +30,19 @@ const AlphaValueInput = ({
           onChangeCustomTheme(propertyName, val);
         }}
       />
-    </AlphaValueInputContainer>
+    </DxcFlex>
   );
 };
 
-const AlphaValueInputContainer = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
 const StyledInput = styled.input`
   font: normal 12px/17px Open Sans;
-  width: 80px;
+  width: 88px;
+
+  &:focus {
+    border-color: transparent;
+    border-radius: 2px;
+    outline: 2px solid #0095ff;
+  }
 `;
 
 export default AlphaValueInput;
