+++ website/screens/theme-generator/components/widgets/BorderStyleInput.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styled from "styled-components";
+import ThemeInputWidgetProps from "./types";
 
 const styleOptions = [
   "none",
@@ -18,7 +19,7 @@ const BorderStyleInput = ({
   propertyName,
   propertyValue,
   onChangeCustomTheme,
-}) => {
+}: ThemeInputWidgetProps): JSX.Element => {
   return (
     <StyledSelect
       onChange={(event) => {
@@ -41,7 +42,13 @@ const BorderStyleInput = ({
 const StyledSelect = styled.select`
   font: normal 12px/17px Open Sans;
   height: 23px;
-  width: 87px;
+  width: 88px;
+
+  &:focus {
+    border-color: transparent;
+    border-radius: 2px;
+    outline: 2px solid #0095ff;
+  }
 `;
 
 export default BorderStyleInput;
