+++ website/screens/theme-generator/components/widgets/types.ts
@@ -1,0 +1,8 @@
+
+type ThemeInputWidgetProps = {
+  propertyName: string;
+  propertyValue: string;
+  onChangeCustomTheme: (propertyName: string, propertyValue: string) => void;
+};
+
+export default ThemeInputWidgetProps;
